requirements.txt
streamlit
ultralytics
pillow
numpy
reportlab
openai
huggingface_hub
